backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
Pillow==9.5.0
channels==4.0.0
channels-redis==4.1.0
